Переносит вспомогательную функцию html из ComponentView в utils.js
