+++ src/view/editor-view.js
@@ -1,13 +1,13 @@
 /** @typedef {import('./point-view').default} PointView */
 
-import ComponentView, { html } from './component-view.js';
+import ComponentView from './component-view.js';
 import TypeSelectView from './type-select-view.js';
 import DestinationSelectView from './destination-select-view.js';
 import OfferSelectView from './offer-select-view.js';
 import DestinationDetailsView from './destination-details-view.js';
 import PriceInputView from './price-input-view.js';
 import DatePickerView from './date-picker-view.js';
-import { isKeyEscape } from '../utils.js';
+import { html, isKeyEscape } from '../utils.js';
 
 export default class EditorView extends ComponentView {
   #linked = null;
