+++ src/view/filter-select-view.js
@@ -1,5 +1,6 @@
-import ComponentView, { html } from './component-view.js';
+import ComponentView from './component-view.js';
 import FilterOptionView from './filter-option-view.js';
+import { html } from '../utils.js';
 
 export default class FilterSelectView extends ComponentView {
   /**
