+++ src/view/point-offer-view.js
@@ -1,4 +1,5 @@
-import ComponentView, { html } from './component-view.js';
+import ComponentView from './component-view.js';
+import { html } from '../utils.js';
 
 export default class PointOfferView extends ComponentView {
   constructor() {
